Moving fetchProfileInfo, updateProfileInfo, and updateActivity functions into profile-helper from session-core; Moving unit test for fetchProfileInfo from session-core.test.js into profile-helper.test.js
